+++ hive/lib/src/annotations/generate_adapters.dart
@@ -1,3 +1,4 @@
+import 'package:hive_ce/src/annotations/hive_converter.dart';
 import 'package:meta/meta.dart';
 
 /// Annotation to generate TypeAdapters for the given [specs]
@@ -9,6 +10,7 @@ class GenerateAdapters {
     this.specs, {
     this.firstTypeId = 0,
     this.reservedTypeIds = const {},
+    this.converters = const [],
   });
   // coverage:ignore-end
 
@@ -22,6 +24,12 @@ class GenerateAdapters {
   ///
   /// These type ids will be skipped during generation
   final Set<int> reservedTypeIds;
+
+  /// A list of [HiveConverter]s to apply when generating adapters for [specs]
+  ///
+  /// A converter is selected by matching its type parameter [T] against each
+  /// field type. See [HiveConverter] for details.
+  final List<HiveConverter> converters;
 }
 
 /// Configuration that specifies the generation of a TypeAdapter