+++ hive_generator/example/lib/hive/hive_adapters.g.dart
@@ -157,6 +157,49 @@ class ClassSpec3Adapter extends TypeAdapter<ClassSpec3> {
           typeId == other.typeId;
 }
 
+class ClassSpec5Adapter extends TypeAdapter<ClassSpec5> {
+  @override
+  final typeId = 55;
+
+  @override
+  ClassSpec5 read(BinaryReader reader) {
+    final numOfFields = reader.readByte();
+    final fields = <int, dynamic>{
+      for (int i = 0; i < numOfFields; i++) reader.readByte(): reader.read(),
+    };
+    return ClassSpec5(
+      const UriConverter().fromHive(fields[0] as String),
+      fields[1] == null
+          ? null
+          : const UriConverter().fromHive(fields[1] as String),
+    );
+  }
+
+  @override
+  void write(BinaryWriter writer, ClassSpec5 obj) {
+    writer
+      ..writeByte(2)
+      ..writeByte(0)
+      ..write(const UriConverter().toHive(obj.url))
+      ..writeByte(1)
+      ..write(
+        obj.optionalUrl == null
+            ? null
+            : const UriConverter().toHive(obj.optionalUrl as Uri),
+      );
+  }
+
+  @override
+  int get hashCode => typeId.hashCode;
+
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      other is ClassSpec5Adapter &&
+          runtimeType == other.runtimeType &&
+          typeId == other.typeId;
+}
+
 class ClassSpec4Adapter extends TypeAdapter<ClassSpec4> {
   @override
   final typeId = 54;